tests/custom-html5lib-tests/tree-construction/custom.dat

#data
<script></code><a href=foo></a></script>
#errors
#document
| <html>
|   <head>
|     <script>
|       "</code><a href=foo></a>"
|   <body>
#errors
